backend/requirements.txt
fastapi
uvicorn 
python-dotenv
python-docx
sentence-transformers
numpy
faiss-cpu
PyMuPDF
langchain
pandas
bs4
pinecone-client
